# Vectorize `basic_string::rfind` (the single character overload)