@StephanTLavavej 2024-11-19
5950X results:

Benchmark                                                               | Before  |   After | Speedup
------------------------------------------------------------------------|--------:|--------:|-------:
`bm<char, not_highly_aligned_allocator, Op::StringRFind>/8021/3056`     |  656 ns | 49.1 ns | 13.36
`bm<char, not_highly_aligned_allocator, Op::StringRFind>/63/62`         | 16.5 ns | 5.33 ns |  3.10
`bm<char, not_highly_aligned_allocator, Op::StringRFind>/31/30`         | 9.69 ns | 8.72 ns |  1.11
`bm<char, not_highly_aligned_allocator, Op::StringRFind>/15/14`         | 5.50 ns | 8.25 ns |  0.67
`bm<char, not_highly_aligned_allocator, Op::StringRFind>/7/6`           | 3.01 ns | 5.68 ns |  0.53
`bm<char, highly_aligned_allocator, Op::StringRFind>/8021/3056`         |  652 ns | 48.9 ns | 13.33
`bm<char, highly_aligned_allocator, Op::StringRFind>/63/62`             | 16.4 ns | 5.29 ns |  3.10
`bm<char, highly_aligned_allocator, Op::StringRFind>/31/30`             | 9.65 ns | 8.51 ns |  1.13
`bm<char, highly_aligned_allocator, Op::StringRFind>/15/14`             | 5.49 ns | 8.33 ns |  0.66
`bm<char, highly_aligned_allocator, Op::StringRFind>/7/6`               | 3.04 ns | 5.68 ns |  0.54
`bm<wchar_t, not_highly_aligned_allocator, Op::StringRFind>/8021/3056`  |  688 ns | 87.0 ns |  7.91
`bm<wchar_t, not_highly_aligned_allocator, Op::StringRFind>/63/62`      | 16.2 ns | 5.75 ns |  2.82
`bm<wchar_t, not_highly_aligned_allocator, Op::StringRFind>/31/30`      | 9.23 ns | 5.32 ns |  1.73
`bm<wchar_t, not_highly_aligned_allocator, Op::StringRFind>/15/14`      | 5.48 ns | 7.88 ns |  0.70
`bm<wchar_t, not_highly_aligned_allocator, Op::StringRFind>/7/6`        | 2.93 ns | 7.62 ns |  0.38
`bm<char32_t, not_highly_aligned_allocator, Op::StringRFind>/8021/3056` |  663 ns |  172 ns |  3.85
`bm<char32_t, not_highly_aligned_allocator, Op::StringRFind>/63/62`     | 15.7 ns | 6.61 ns |  2.38
`bm<char32_t, not_highly_aligned_allocator, Op::StringRFind>/31/30`     | 8.94 ns | 5.14 ns |  1.74
`bm<char32_t, not_highly_aligned_allocator, Op::StringRFind>/15/14`     | 5.39 ns | 4.69 ns |  1.15
`bm<char32_t, not_highly_aligned_allocator, Op::StringRFind>/7/6`       | 2.92 ns | 4.89 ns |  0.60

Looks good, the only slowdowns are where the character is found *immediately* and the call is super fast anyways.