@StephanTLavavej 2024-11-19
I'm mirroring this to the MSVC-internal repo - please notify me if any further changes are pushed.